betty

+++ 0x1E-search_algorithms/search_algos.h
@@ -4,6 +4,6 @@
 
 int linear_search(int *array, size_t size, int value);
 int binary_search(int *array, size_t size, int value);
-int search(size_t l,size_t r,int *array,int value);
+int search(int l,int r,int *array,int value);
 
 #endif
